requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
gunicorn
psycopg2
flask_jwt_extended
opencv-python
pillow
moviepy
numpy
